+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](semver).
 ### Added
 - Allow CORS during local development. [#105][#105]
 - Add BUILD_ENV environment variable. [#105][#105]
+- Cache dashboard pledge data client-side. [#104][#104]
 
 ### Changed
 - Update resource links during local development. [#105][#105]
@@ -19,6 +20,7 @@ and this project adheres to [Semantic Versioning](semver).
 ### Removed
 
 ### Fixed
+- Allow auth requests from both API URL and site URL. [#104][#104]
 
 ### Security
 - Bump `bcrypt` a major version.
@@ -108,4 +110,5 @@ and this project adheres to [Semantic Versioning](semver).
 [#89]: https://github.com/Visual-Communications/fair-housing-pledge/issues/89
 [#92]: https://github.com/Visual-Communications/fair-housing-pledge/issues/92
 [#93]: https://github.com/Visual-Communications/fair-housing-pledge/issues/93
+[#104]: https://github.com/Visual-Communications/fair-housing-pledge/issues/104
 [#105]: https://github.com/Visual-Communications/fair-housing-pledge/issues/105